+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## TBD
+### Fixed
+- `apiary-metastore-listener` now serializes metastore table properties as a Map instead of a String. 
+
 ## [3.0.0] - 2019-04-12
 ### Changed
 - Maven group ID is now `com.expediagroup.apiary` (was `com.expedia.apiary`).